Split CLI commands to several files for easier editing and naming.
